infra: don't pass external domain to Django
This sets ALLOWED_HOSTS to ["*"] in order for the ALB healthcheck to be able to
make requests to the app.

The ALB already enforces that the right host header is sent, and also the ALB
enforces that requests can only be sent from our CloudFront distribution, which
itself enforces that the right host header is sent, so there are enough layers
of projection on the host header, so it should be safe to allow all hosts from
Django's point of view.