refactor: tidy search code and make reusable
There isn't a current plan to reuse the search code in multiple places.
However, there is a plan to move the transcription code to the front page, and
in multiple places. Since that's chunkier, doing this smaller version version
as a setting stone/POC to see how it feels and works.