chore: remove code only used for local development of transcription
This removes code that is only in place for local development of the
transcription feature, and specifically removes the code that returns
credentials that are not generated on demand for each user/page load.

This is so the same code path is used locally as when deployed to ECS, which
both ensures that if it works locally, it will work when deployed, and makes it
easier to reason about what is happening in terms of the credentials flow.

This change also adds a brief explanation to the README on how the credentials
and transcription flow works, and includes detailed instructions on how to have
transcription working locally in an as-similar-as-possible-way to when deployed
in ECS.