Refactor/rename refs to access group (#359)
Currently some of the UI is showing to users when it shouldn't. This is
because some refs to a user group is incorrect and needs renaming from
"Viewer" to "Viewer access". To make sure we stay consistent I have
reworked the string values to refer to constants in the settings.py.