Styling the one page document (#381)
This change applies some basic print styles for when the user wants to
print the one page document. Most of the style rules are scoped to the
company briefing page for this purpose only. I have also hidden the
"print" button from the "view all engagements" page as its only needed
for the company briefing page.